Removed routing from WASM
