+++ packages/rewild-assembly/lib/index.ts
@@ -1,15 +1,6 @@
 export * from './extras/io';
 export * from './extras/ui';
-export {
-  getRuntime,
-  init,
-  resize,
-  update,
-  addNodeToRuntime,
-  removeNodeFromRuntime,
-  getScene,
-  getCamera,
-} from './objects/routing/AsSceneManager';
+export * from './objects';
 export { createMeshPipelineInstance } from './pipelines/MeshPipelineInstance';
 export { addPipelineAttribute } from './pipelines/PipelineInstance';
 export { createMeshComponent } from './components/MeshComponent';
@@ -18,24 +9,19 @@ export * from './components/index';
 export * from './objects/physics/index';
 
 export {
-  createContainer,
-  addAsset,
-  addChildNode,
-  removeChildNode,
-  createLevel,
-  createLink,
-  connectLink,
-  getNodePortal,
-  addNodePortal,
-  removeNodePortal,
-  createPortal,
-} from './objects/routing';
+  getRuntime,
+  init,
+  resize,
+  update,
+  getScene,
+  getCamera,
+} from './WasmManager';
+
 export {
   getCameraProjectionInverseMatrix,
   getCameraProjectionMatrix,
   getCameraWorldInverseMatrix,
 } from './cameras/Camera';
-export * from './objects';
 
 export {
   addChild,
