--- packages/rewild-assembly/lib/objects/routing/core/Node.ts
@@ -1,126 +1,0 @@
-import { Runtime } from ".";
-import { Portal } from "./Portal";
-
-export class Node {
-  protected portals: Portal[];
-  children: Node[];
-  protected parent: Node | null;
-
-  runtime: Runtime | null;
-  initialized: boolean;
-  mounted: boolean;
-  autoDispose: boolean;
-  isDisposed: boolean;
-  name: string;
-
-  constructor(name: string, autoDispose: boolean = true) {
-    this.name = name;
-    this.children = [];
-    this.portals = [];
-    this.initialized = false;
-    this.isDisposed = false;
-    this.parent = null;
-    this.mounted = false;
-    this.autoDispose = autoDispose;
-  }
-
-  dispose(): void {
-    this.isDisposed = true;
-
-    for (let p: i32 = 0, pl: i32 = this.portals.length; p < pl; p++) {
-      const portal = unchecked(this.portals[p]);
-      portal.dispose();
-    }
-
-    for (let i: i32 = 0, l: i32 = this.children.length; i < l; i++) {
-      const child = unchecked(this.children[i]);
-      child.dispose();
-    }
-
-    this.parent = null;
-    this.runtime = null;
-  }
-
-  addChild(node: Node): Node {
-    if (node.parent) {
-      node.parent!.removeChild(node);
-    }
-
-    if (this.children.indexOf(node) == -1) {
-      this.children.push(node);
-      node.runtime = this.runtime;
-    }
-
-    node.parent = this;
-    return node;
-  }
-
-  removeChild(node: Node): Node {
-    const i = this.children.indexOf(node);
-    if (i != -1) {
-      this.children.splice(i, 1);
-      node.parent = null;
-    }
-    return node;
-  }
-
-  addPortal(portal: Portal): Portal {
-    if (this.portals.indexOf(portal) == -1) {
-      this.portals.push(portal);
-      portal.node = this;
-    }
-    return portal;
-  }
-
-  removePortal(portal: Portal): Portal {
-    const i = this.portals.indexOf(portal);
-    if (i != -1) {
-      this.portals.splice(i, 1);
-    }
-    return portal;
-  }
-
-  init(): void {
-    this.initialized = true;
-  }
-
-  getPortal(name: string): Portal | null {
-    const portals = this.portals;
-    for (let i: i32 = 0, l = portals.length; i < l; i++)
-      if (portals[i].name == name) return portals[i];
-
-    return null;
-  }
-
-  mount(): void {
-    this.mounted = true;
-  }
-
-  unMount(): void {
-    this.mounted = false;
-  }
-
-  onUpdate(delta: f32, total: u32): void {}
-
-  enter(portalEntered: Portal): void {}
-}
-
-export function addChildNode(parent: Node, child: Node): void {
-  parent.addChild(child);
-}
-
-export function removeChildNode(parent: Node, child: Node): void {
-  parent.removeChild(child);
-}
-
-export function getNodePortal(node: Node, portalName: string): Portal | null {
-  return node.getPortal(portalName);
-}
-
-export function addNodePortal(node: Node, portal: Portal): void {
-  node.addPortal(portal);
-}
-
-export function removeNodePortal(node: Node, portal: Portal): void {
-  node.removePortal(portal);
-}
