--- packages/rewild-assembly/lib/objects/routing/core/index.ts
@@ -1,6 +1,0 @@
-export * from "./Container";
-export * from "./Level";
-export * from "./Node";
-export * from "./Runtime";
-export * from "./Link";
-export * from "./Portal";
