--- packages/rewild-assembly/lib/objects/routing/custom/index.ts
@@ -1,3 +1,0 @@
-export * from "./Level1";
-export * from "./MainMenu";
-export * from "./EditorContainer";
