--- packages/rewild-assembly/lib/tests/routers/index.ts
@@ -1,3 +1,0 @@
-export { createDebugNode } from "./DebugNode";
-export { createDebugContainer } from "./DebugContainer";
-export { createDebugLevel } from "./DebugLevel";
