--- packages/rewild-common/lib/ContainerType.ts
@@ -1,6 +1,0 @@
-export class ContainerTypes {
-  static Editor: string = "Editor";
-  static Level1: string = "Level1";
-  static MainMenu: string = "MainMenu";
-  static Default: string = "Default";
-}
